Save... lenovo makes shitty laptops that can't be trusted.
