+++ imports/api/index.js
@@ -1,5 +1,10 @@
+import { Meteor } from 'meteor/meteor';
 
-import './device/methods.js';
+if (Meteor.isServer) {
+	import './thing/server-methods.js';
+}
+
+// 
 import './environment/methods.js';
 import './notifications/methods.js';
 import './thing/methods.js';
